+++ bootstrap/ssr/assets/Index-BNgg2TIi.js
@@ -1,0 +1,60 @@
+import { jsx } from "react/jsx-runtime";
+import "react";
+import { M as MobileLayout } from "./MobileLayout-kgatcMk_.js";
+import NewNasabah from "./NewNasabah-yQyYQS-j.js";
+import "./AuthenticatedLayout-DhCvM_RZ.js";
+import "@inertiajs/react";
+import "./command-int9mZp7.js";
+import "@radix-ui/react-slot";
+import "class-variance-authority";
+import "./utils-DzFuPzol.js";
+import "clsx";
+import "tailwind-merge";
+import "dayjs";
+import "./button-MTjEwktD.js";
+import "./input-BHD-__le.js";
+import "@radix-ui/react-separator";
+import "@radix-ui/react-dialog";
+import "@radix-ui/react-icons";
+import "@radix-ui/react-tooltip";
+import "@radix-ui/react-dropdown-menu";
+import "cmdk";
+import "lucide-react";
+import "./dialog-8a8NNlps.js";
+import "./popover-B9xSK2Gy.js";
+import "@radix-ui/react-popover";
+import "sweetalert2";
+import "sweetalert2-react-content";
+import "./Loading-DPcVnbEk.js";
+import "react-dom";
+import "@headlessui/react";
+import "./InputError-cRVTeK4i.js";
+import "./card-WipCscFv.js";
+import "./tabs-BAdW7Ajv.js";
+import "@radix-ui/react-tabs";
+import "./label-F4MKz6SO.js";
+import "@radix-ui/react-label";
+import "axios";
+import "react-currency-input-field";
+import "./SelectComponent-DUhN-d41.js";
+import "./RiwayatPengajuan-BgwqMR0N.js";
+import "./table-Dwx5kZ1B.js";
+import "@tanstack/react-table";
+import "./BadgeStatus-Cj6LHx_M.js";
+import "./badge-CBTu05xj.js";
+import "./Checkbox-DLnjqb3e.js";
+import "./FormatNumbering-CTESXtjE.js";
+import "react-number-format";
+import "./NoEditOverlay-GIB1h_zq.js";
+const Index = () => {
+  return /* @__PURE__ */ jsx(MobileLayout, { children: /* @__PURE__ */ jsx(
+    NewNasabah,
+    {
+      onClosed: () => void 0,
+      submitUrl: route("transaction.store_buku_transaksi")
+    }
+  ) });
+};
+export {
+  Index as default
+};